requirements.txt
atheris==3.0.0
atomicwrites==1.4.1
BeautifulSoup==3.2.2
beautifulsoup4==4.14.3
brotli==1.2.0
brotlicffi==1.2.0.0
cached_property==2.0.1
cffi==2.0.0
ConfigParser==7.2.0
cython==3.2.4
docutils==0.22.4
github_copilot_sdk==0.1.15
grpc_reflection==1.0.0
grpc_status==1.0.0
grpc_tools==1.0.0
h2==4.3.0
html5lib==1.1
HTMLParser==0.0.2
hypothesis==6.150.2
importlib_metadata==8.7.1
interegular==0.3.3
ipython==9.9.0
ipywidgets==8.1.8
jaraco.util==15.0.0
jnius==1.1.0
js==1.0
js2py==0.74
jsonpath_ng==1.7.0
keyring==25.7.0
lxml_html_clean==0.4.3
mod==0.3.0
oauth2client==4.1.3
pydot==4.0.1
PyInstaller==6.18.0
pyodide==0.0.2
pyperf==2.9.0
pytest==8.4.2
python_bcrypt==0.3.2
pyu2f==0.1.5
redis==7.1.0
regex==2026.1.15
rfc3987==1.3.8
scanoss_winnowing==0.7.0
setuptools==80.10.1
simplejson==3.20.2
Sphinx==9.1.0
thread==2.0.6
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zipp==3.23.0
